add logic to policy and controller for authorisation and authentication

+++ app/Http/Controllers/PostController.php
@@ -3,10 +3,20 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Routing\Controllers\Middleware;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
-class PostController extends Controller
+class PostController extends Controller implements HasMiddleware
 {
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth', except: ['index', 'show']),
+        ];
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -57,6 +67,9 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        // Authorisation
+        Gate::authorize('modify', $post);
+
         return view('posts.edit', ['post' => $post]);
     }
 
@@ -65,6 +78,9 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        // Authorisation
+        Gate::authorize('modify', $post);
+
         // Validate
 
         $fields = $request->validate([
@@ -85,6 +101,9 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        // Authorisation
+        Gate::authorize('modify', $post);
+
         // Delete the post
         $post->delete();
